nanda_adapter/agent/requirements.txt
# Core dependencies for megabrain.py
langchain-core
langchain-anthropic
python-dotenv
requests

# NANDA Adapter
nanda-adapter
